bot/requirements.txt
appdirs==1.4.4
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cryptography==3.4.8
cssselect==1.1.0
fake-useragent==0.1.11
Flask==2.0.1
greenlet==1.1.1
h11==0.12.0
h2==4.0.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
kaitaistruct==0.9
lxml==4.6.3
MarkupSafe==2.0.1
mysql==0.0.3
mysql-client==0.0.1
mysql-connector==2.2.9
mysqlclient==2.0.3
parse==1.19.0
pyasn1==0.4.8
pycparser==2.20
pydivert==2.1.0
pyee==8.2.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyppeteer==0.2.6
pyquery==1.4.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
tqdm==4.62.2
urllib3==1.26.6
w3lib==1.22.0
websockets==9.1
Werkzeug==2.0.1
wsproto==1.0.0
zipp==3.5.0